MINOR: mt_list: Implement MT_LIST_POP_LOCKED()
Implement MT_LIST_POP_LOCKED(), that behaves as MT_LIST_POP() and
removes the first element from the list, if any, but keeps it locked.

This should be backported to 3.2, as it will be use in a bug fix in the
stick tables that affects 3.2 too.

(cherry picked from commit cf26745857d1c4bbaf6404f57c47bb7c114c6905)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>