MINOR: sched: let's permit to share the local ctx between threads
The watchdog timer has to go through complex operations due to not being
able to check if another thread's scheduler is still ticking. This is
simply because the scheduler status is marked as thread-local while it
could in fact also be an array. Let's do that (and align the array to
avoid false sharing) so that it's now possible to check any scheduler's
status.

(cherry picked from commit 7c7e17a6057652d6702c114ab6c4ba8cd301d54b)
[wt: backported for the next wdt fix]
Signed-off-by: Willy Tarreau <w@1wt.eu>