MEDIUM: servers: Schedule the server requeue target on creation
On creation, schedule the server requeue once it's been created.
It is possible that when the server went up, it tried to queue itself
into the lb specific code, failed to do so, and expect the tasklet to
run to take care of that.

This should be backported to 3.2.
This is part of an attempt to fix github issue #3143.

(cherry picked from commit 822ee90dc2b5910a88948053ffa21b4caab53d28)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>