BUG/MINOR: quic: check applet_putchk() for 'show quic' first line
Ensure applet_putchk() return value is checked when outputing via the
CLI 'show quic' header line.

This is only to align with other usages of the same function, as trash
output buffer should always be large enough for it. As such, the command
is simply aborted if this is not the case.

This should fix coverity report from github issue #3139.

This could be backported up to 2.8.

(cherry picked from commit 5067a15870ca27bb6519f9076b8c8a7c7db8ab14)
Signed-off-by: Willy Tarreau <w@1wt.eu>