MINOR: ncbuf: extract common types
ncbuf is a module which provide a non-contiguous buffer type
implementation. This patch extracts some basic types related to it into
a new file ncbuf_common.h.

This patch will be useful to provide a new non-contiguous buffer
alternative implementation based on a bitmap.

This patch is not a bug fix. However, it is necessary for ncbmbuf
implementation which will be required to fix a QUIC issue on CRYPTO
frames parsing. This, it will be necessary to backport the current patch
prior to the fix to come.

(cherry picked from commit 59f0bafef28259f4bceebb00bbce4c48e5dd5791)
[wt: backported as needed by a subsequent fix]
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 98ddaacf54321f7bc5348f0f217b4e64c889f6aa)
Signed-off-by: Willy Tarreau <w@1wt.eu>