MINOR: ncbmbuf: implement advance operation
Implement ncbmb_advance() function for the ncbmbuf type. This allows to
remove bytes in front of the buffer, regardless of the existing gaps.
This is implemented by resetting the corresponding bits of the bitmap.

As the previous patch, this commit must be backported prior to the fix
to come on QUIC CRYPTO frames parsing.

(cherry picked from commit 8b8ab2824e7cc74fa2d93b5d71ca1aaec7925e15)
[wt: backported as needed by a subsequent fix]
Signed-off-by: Willy Tarreau <w@1wt.eu>