BUG/MEDIUM: mt_list: Make sure not to unlock the element twice
In mt_list_delete(), if the element was not in a list, then n and p will
point to it, and so setting n->prev and n->next will be enough to unlock it.
Don't do it twice, as once it's been done the first time, another thread may
be working with it, and may have added it to a list already, and doing it
a second time can lead to list inconsistencies.

This should be backported up to 2.8.

(cherry picked from commit 7a33b90b3cdcf26e4a73bee9e89a4fe670eee167)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 068fdde244d9eb113fa3e8a107fbdaada4fe04c4)
Signed-off-by: Willy Tarreau <w@1wt.eu>