BUG/MEDIUM: h1-htx: Don't set HTX_FL_EOM flag on 1xx informational messages
1xx informational messages are part of the HTTP response. It is not expected
to have a HX_FL_EOM flag set after parsing such messages when received from
a server. It is espacially important whne an informational messages is
processed on client side while the final response was not recieved yet, to
not erroneously detect the end of the message.

The HTTP multiplexers seem to ignore the HTX_FL_EOM flag for information
messages, but it remains an error from the HTX specification point of
view. So it must be fixed.

While it should theorically be backported as far as 3.0, it is a good idea
to not do so for now because no bug was reported and regressions may happen.

(cherry picked from commit 18ece2b4249956cb759552ff831d3822b664f223)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit ccb47bb83638b5e6a5f935e25a2c22f59e5af1dc)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>